[BUGFIX] Do not crash when the TYPO3 cache is empty and flux is used
When using typo3_encore in combination with flux, an exception occurs
when loading a page after freshly clearing all caches:

> The build "_default" is not configured

This happens because with an fully empty cache, TYPO3 needs to re-build the
TypoScript configuration.
When flux is used, building the TS array means parsing all content elements,
which in turn instantiates <encore:renderWebpackScriptTags> tags in them.

Because of constructor dependency injection,
EntrypointLookupFactory::getCollection() gets called, which fails badly.

After fixing this, the settings are still empty because
- again due to constructor dependency injection - AssetRegistry
fetches the (still empty) settings, which get cached in SettingsService.
When the empty settings are used later, the build _default is still missing.

Resolves: https://github.com/sabbelasichon/typo3_encore/issues/166